Move statement tree
