+++ projects/batfish/src/main/java/org/batfish/grammar/hierarchical/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Utilities to pre-process vendors that have hierarchical config and allow operations on the entire
+ * subtrees of the configuration (e.g., via a delete command)
+ */
+package org.batfish.grammar.hierarchical;
